# frontend/src/lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FlowCyt Panel API",
    "description": "Hybrid AI tool for multi-color flow cytometry panel design.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/quality-registry/issues": {
      "post": {
        "tags": [
          "quality-registry"
        ],
        "summary": "Create Issue",
        "description": "Create a new quality issue and trigger projection update.",
        "operationId": "create_issue_api_v1_quality_registry_issues_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QualityIssueCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QualityIssueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quality-registry/candidates/lookup": {
      "post": {
        "tags": [
          "quality-registry"
        ],
        "summary": "Candidate Lookup",
        "description": "Look up candidate antibody matches from inventory.\n\nSimple heuristic: filter inventory by marker and fluorochrome,\nrank by exactness of match.",
        "operationId": "candidate_lookup_api_v1_quality_registry_candidates_lookup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateLookupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateLookupResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quality-registry/candidates/confirm": {
      "post": {
        "tags": [
          "quality-registry"
        ],
        "summary": "Candidate Confirm",
        "description": "Confirm a candidate entity selection for a quality issue.",
        "operationId": "candidate_confirm_api_v1_quality_registry_candidates_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateConfirmWithIssue"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QualityIssueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/quality-registry/issues": {
      "get": {
        "tags": [
          "admin",
          "admin-quality-registry"
        ],
        "summary": "List Issues",
        "description": "List quality issues, optionally filtered by status.",
        "operationId": "list_issues_api_v1_admin_quality_registry_issues_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QualityIssueResponse"
                  },
                  "title": "Response List Issues Api V1 Admin Quality Registry Issues Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CandidateConfirmWithIssue": {
        "properties": {
          "issue_id": {
            "type": "string",
            "minLength": 1,
            "title": "Issue Id"
          },
          "entity_key": {
            "$ref": "#/components/schemas/EntityKey"
          }
        },
        "type": "object",
        "required": [
          "issue_id",
          "entity_key"
        ],
        "title": "CandidateConfirmWithIssue",
        "description": "Confirm a candidate entity selection for a specific issue."
      },
      "CandidateLookupRequest": {
        "properties": {
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text",
            "description": "Natural-language description of the antibody."
          },
          "species": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Species"
          },
          "marker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Marker"
          },
          "fluorochrome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fluorochrome"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "CandidateLookupRequest",
        "description": "Request to look up candidate antibody matches."
      },
      "CandidateLookupResponse": {
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/CandidateMatch"
            },
            "type": "array",
            "title": "Candidates"
          }
        },
        "type": "object",
        "title": "CandidateLookupResponse",
        "description": "Response containing candidate antibody matches."
      },
      "CandidateMatch": {
        "properties": {
          "entity_key": {
            "$ref": "#/components/schemas/EntityKey"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "matched_marker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matched Marker"
          }
        },
        "type": "object",
        "required": [
          "entity_key",
          "confidence",
          "source"
        ],
        "title": "CandidateMatch",
        "description": "A single candidate match from the lookup."
      },
...
      "EntityKey": {
        "properties": {
          "species": {
            "type": "string",
            "title": "Species"
          },
          "normalized_marker": {
            "type": "string",
            "title": "Normalized Marker"
          },
          "clone": {
            "type": "string",
            "title": "Clone"
          },
          "brand": {
            "type": "string",
            "title": "Brand"
          },
          "catalog_number": {
            "type": "string",
            "title": "Catalog Number"
          },
          "lot_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lot Number"
          }
        },
        "type": "object",
        "required": [
          "species",
          "normalized_marker",
          "clone",
          "brand",
          "catalog_number"
        ],
        "title": "EntityKey",
        "description": "Canonical identity for a specific antibody reagent.\n\nEquality & hashing are based on the 5 required fields.\n``lot_number`` is tracked as metadata but is NOT part of identity."
      },
      "FeedbackKey": {
        "properties": {
          "species": {
            "type": "string",
            "title": "Species"
          },
          "normalized_marker": {
            "type": "string",
            "title": "Normalized Marker"
          },
          "fluorochrome": {
            "type": "string",
            "title": "Fluorochrome"
          },
          "brand": {
            "type": "string",
            "title": "Brand"
          },
          "clone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clone"
          }
        },
        "type": "object",
        "required": [
          "species",
          "normalized_marker",
          "fluorochrome",
          "brand"
        ],
        "title": "FeedbackKey",
        "description": "Identity key for a quality-feedback submission.\n\nEquality & hashing are based on all non-None fields, so two submissions\nthat differ only in clone (where one has clone=None) are *not* equal."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QualityIssueCreate": {
        "properties": {
          "issue_text": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Issue Text"
          },
          "reported_by": {
            "type": "string",
            "minLength": 1,
            "title": "Reported By"
          },
          "species": {
            "type": "string",
            "minLength": 1,
            "title": "Species"
          },
          "marker": {
            "type": "string",
            "minLength": 1,
            "title": "Marker"
          },
          "fluorochrome": {
            "type": "string",
            "minLength": 1,
            "title": "Fluorochrome"
          },
          "brand": {
            "type": "string",
            "minLength": 1,
            "title": "Brand"
          },
          "clone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clone"
          }
        },
        "type": "object",
        "required": [
          "issue_text",
          "reported_by",
          "species",
          "marker",
          "fluorochrome",
          "brand"
        ],
        "title": "QualityIssueCreate",
        "description": "Payload for submitting a new quality issue."
      },
      "QualityIssueResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "feedback_key": {
            "$ref": "#/components/schemas/FeedbackKey"
          },
          "entity_key": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EntityKey"
              },
              {
                "type": "null"
              }
            ]
          },
          "issue_text": {
            "type": "string",
            "title": "Issue Text"
          },
          "reported_by": {
            "type": "string",
            "title": "Reported By"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "submitted"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "feedback_key",
          "issue_text",
          "reported_by",
          "created_at",
          "updated_at"
        ],
        "title": "QualityIssueResponse",
        "description": "Full representation of a persisted quality issue."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}